requirements.txt
pyyaml
numpy
pandas
ClauseWizard
matplotlib
seaborn
tqdm
openpyxl